chore: export comment node
